refactor(charm): move AAR configuration out of the charm
This commit refactors the `AARSnap` into an `AAR` object and moves
confiuring the AAR with the certificate setup into this new class. This
increases the sparation of configuration concerns for the snap from the charm.
