+++ config.yaml
@@ -1,27 +1,4 @@
 options:
-  ua_token:
-    default: ""
-    type: string
-    description: |
-      Ubuntu Advantage token you have received with your license of Anbox Cloud
-  ua_source:
-    default: ""
-    type: string
-    description: |
-      APT source to use instead of the automatic one generated from UA.
-      Overrides UA source if both are set.
-  ua_source_key:
-    default: ""
-    type: string
-    description: |
-      Key ID to import to the APT keyring in order to access the specified
-      source archive.
-      Required if ua_source is set.
-  ua_use_staging:
-    default: False
-    type: boolean
-    description: |
-      Use staging contracts API service
   port:
     type: int
     default: 3000
@@ -38,19 +15,3 @@ options:
     type: string
     default: ""
     description: Location the registry is available on. If not set its public address will be used.
-  nagios_context:
-    default: "juju"
-    type: string
-    description: |
-      Used by the nrpe subordinate charms.
-      A string that will be prepended to instance name to set the host name
-      in nagios. So for instance the hostname would be something like:
-        juju-myservice-0
-      If you're running multiple environments with the same services in them
-      this allows you to differentiate between them.
-  nagios_servicegroups:
-    default: ""
-    type: string
-    description: |
-      A comma-separated list of nagios servicegroups.
-      If left empty, the nagios_context will be used as the servicegroup
